support functions for org-specific base templates per domain
We want to be able to support organization-branded versions of
the DDR.  Partners will want to display their collections (and
only theirs) and customize the UI.

settings.ORG_DOMAINS is used to list domains that belong to
specific organizations.  If the request domain matches an
organization domain the org keyword is added to sitewide context.
If org-specific base template exists it also is added.