first pass at django-rest-framework API; detail endpoints
First draft implementation of API version of detail pages.
We're not using the Django ORM so there are no serializers.
We're mostly taking the ES data, adding URLs to it, and passing
it to rest_framework.Response.
Each level lists some of the children -- I'm not yet sure how DRF
handles paging.