context_processors,templates: ASSETS_BASE to avoid mixed-content
Package the HTTP protocol (http or https), the HTTP host (domain
name or IP addr) and assets base into a var for templates.
The purpose of this is to deliver all assets in the same HTTP
protocol so as to not generate mixed content errs when in HTTPS.
Starting assets URLs with double slashes (e.g. "//assets/...")
is supposed to do the same thing but doesn't work in local dev
e.g. with an IP address and no domain name.