docstore,search: Restructure to match ddr-local/ddr-cmdln
This project is simply a Django app, but ddr-local is a Django app
built on a core that does not/can not know anything about Django,
which means no Request objects.
That app needs to make it possible for ddr-local to override
ddr-cmdln code easily.
This change restructures certain parts of the search code to
match ddr-cmdln as closely as possible.