models: Disable SSL cert verification when checking img file size
`ui.models.file_size` gets only an HTTP header for an image to be
used to check the file size.  This is used to check file size of
images on the media server and uses the "outside" URL.
This doesn't work in dev because we're using self-signed certs.
This commit disables the verification so things work in dev.
Should not be a problem in prod since we're using verified certs
and it's just the header (queue ominious music?).